feat: move metamask component to own file
User will now have to click sign in, in order to activate metamask.
